has_children from byte to int

+++ common/src/main/java/cn/malgo/annotation/common/dal/model/Concept.java
@@ -21,7 +21,7 @@ public class Concept {
     private String conceptName;
 
     @Column(name = "has_children")
-    private Byte hasChildren;
+    private int hasChildren;
 
     /**
      * @return concept_id
@@ -96,7 +96,7 @@ public class Concept {
     /**
      * @return has_children
      */
-    public Byte getHasChildren() {
+    public int getHasChildren() {
         return hasChildren;
     }
